rename: auth폴더 정리, 타입 옮김
