+++ src/apis/auth/auth.type.ts
@@ -1,0 +1,6 @@
+export interface FormValues {
+  email: string;
+  nickname?: string;
+  password: string;
+  passwordConfirm?: string;
+}
